Processor thd now waits on a conditional variable. This way the producers now can notify it to begin consumption and the controller can notify it when its time to die. Added description to scheduler task.

+++ recorder/src/main/cpp/ti_thd.cc
@@ -59,7 +59,7 @@ struct ThreadTargetProc {
     void await_stop() {
         std::unique_lock<std::mutex> l(m);
         if (state != State::stopped) {
-            logger->trace("Will now wait for thread '{}' to be stopped, state as of now: {}", name, state);
+            logger->trace("Will now wait for thread '{}' to be stopped, state as of now: {}", name, static_cast<std::uint32_t>(state));
             v.wait(l, [&] { return state == State::stopped; });
         }
     }
